requirements.txt
python-telegram-bot==13.9
python-decouple==3.5
requests==2.26.0
